Logging Temperature
Not much acutal code has changed since the last commit, but the
fill_zeros funciton in diskio.c now fills 'a's. We need to figure out
how to correctly terminate (and maybe initiate) files.

The main function is now a working temperature logger, and the code has
been tested on the new hardware.